PopUp Fenster überarbeitet, ConfirmationBox vor dem Löschen von Objekten implementiert, Datum-Bug behoben, Schönheitsfehler korrigiert
Aktionen in PopUp Fenstern können jetzt durch Button abgebrochen werden.
User wird nun zur Bestätigung aufgefordert, bevor etwas gelöscht wird.
In expired lists wurden uncompletete Events wegen eines Schreibfehlers im Code nicht angezeigt, das wurde behoben.
PopUp Fenster wurden optisch angepasst ebenso wie andere kleine GUI-Anpassungen.

+++ src/main/java/de/hdmstuttgart/mi/bucketlist/Gui/Controller/PopUpController/ModifyEventController.java
@@ -1,6 +1,7 @@
 package de.hdmstuttgart.mi.bucketlist.Gui.Controller.PopUpController;
 
 import de.hdmstuttgart.mi.bucketlist.Exceptions.ElementAlreadyExistsException;
+import de.hdmstuttgart.mi.bucketlist.Gui.Controller.SceneController.PaneLoader;
 import de.hdmstuttgart.mi.bucketlist.Model.Eventlist;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -39,11 +40,13 @@ public class ModifyEventController {
         stage.close();
     }
 
-    @FXML
+    @FXML //todo rename to deleteEvent
     void deleteList() {
-        this.eventlist.deleteEvent(this.eventName);
-        Stage stage = (Stage) this.backButton.getScene().getWindow();
-        stage.close();
+        if(PaneLoader.loadConfirmationWindow("Are you sure you want to delete the event \" " + this.eventName + " \" ?")){
+            this.eventlist.deleteEvent(this.eventName);
+            Stage stage = (Stage) this.backButton.getScene().getWindow();
+            stage.close();
+        }
     }
 
     @FXML
